requirements.txt
asgiref==3.4.1
Django==3.2.6
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
phonenumbers==8.12.31
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1